Adding basic .obj face parsing
The .obj files polygons are now parsed, there is still some features that needs
to be implemented for the parser to be complete. Indeed the parser can not parse
polygons that have more than 3 vertex, so a triangulation algorithm such as the
ear clipping method needs to be implemented. The parser could also optimize the
vertex number in the array by taking advantage of the indices array.